Create Access private fields using getter and setter methods
Here, we have private variables named age and name. Here, we are trying to access the private variables from other class named Main.

We have used the getter and setter method to access the private variables. Here,

the setter methods setAge() and setName() initializes the private variables
the getter methods getAge() and getName() returns the value of private variables